requirements.txt
asgiref==3.4.1
cloudinary==1.28.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.9
django-taggit==2.0.0
gunicorn==20.1.0
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
